Rename follows_update_spec to follows_bulk_update_spec and update
